feat(dangerouslyIgnoredRequiredPlugins): allow users to disable check for required plugins
This change is motivated by the need to inject an sdk instance that may not have certain plugins available to it. For example, plugin-phone is being deprecated but is still required for the widgets. This approach will allow for users to "dangerously" ignore such plugins when implementing the widget without introducing potential breaking changes to legacy instances of the widget.